[core] Add support for job traits with ID > 128
* Change CTrait internal type ID to uint16
* Modify appropriate getters/setters to match
* Update charentity & ability packet to support 2 more bytes of trait bits
* Add TRAIT_WS_DAMAGE_BOOST to trait.h
* Add "ROV" content tag to spell loading
* This commit enables TRAIT_WS_DAMAGE_BOOST due to the ID no longer wrapping

+++ src/map/lua/lua_baseentity.cpp
@@ -13263,7 +13263,7 @@ void CLuaBaseEntity::instantiateMob(uint32 groupID)
  *  Notes   :
  ************************************************************************/
 
-bool CLuaBaseEntity::hasTrait(uint8 traitID)
+bool CLuaBaseEntity::hasTrait(uint16 traitID)
 {
     if (m_PBaseEntity->objtype != TYPE_PC)
     {
